requirements.txt
pandas>=1.5.0
pymysql>=1.0.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
requests>=2.28.0
mcp>=1.0.0
pydantic>=2.0.0 